compte

+++ ProjetAstroSpringBoot/src/main/java/SoprAjc/ProjetAstroSpringBoot/model/Compte.java
@@ -18,7 +18,7 @@ public class Compte {
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seqLogin")
     @JsonView(JsonViews.Common.class)
-	private Long id;
+	private Integer id;
 	@Column(name = "login", length = 150, unique = true, nullable = false)
 	@NotEmpty
     @JsonView(JsonViews.Common.class)
@@ -41,14 +41,16 @@ public class Compte {
 		this.password = password;
 	}
 	
-	public Long getId() {
+
+	
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
-	
+
 	public String getLogin() {
 		return login;
 	}
